+++ WHATSNEW.md
@@ -21,7 +21,7 @@ Version 3.5 (unreleased)
 
 * New: when creating a Content Type Group, it is possible to set a custom creation date
 
-* New: it is now possible to generate migrations for Sections and Content Type Groups
+* New: it is now possible to generate migrations for Sections, Object State Groups and Content Type Groups
 
 * Fix: content creation from the `generate` command would fail if a field of type Relation has no value